refactor: lazy load shop menu
- It will load when the player opens a category and remain saved until they leave the shop.
