+++ client/main.lua
@@ -194,44 +194,7 @@ end
 ---@param category string
 ---@param targetVehicle number
 local function openVehCatsMenu(category, targetVehicle)
-    local vehMenu = {}
-
-    local function insertVehicle(data)
-        vehMenu[#vehMenu + 1] = {
-            title = ('%s %s'):format(data.brand, data.name),
-            description = locale('menus.veh_price')..lib.math.groupdigits(data.price),
-            serverEvent = 'qbx_vehicleshop:server:swapVehicle',
-            args = {
-                toVehicle = data.model,
-                targetVehicle = targetVehicle,
-                closestShop = insideShop
-            }
-        }
-    end
-
-    for k, vehicle in pairs(VEHICLES) do
-        if vehicle.category == category then
-            local vehicleShop = config.models[k] or config.categories[vehicle.category] or config.default
-            if not vehicleShop then
-                lib.print.debug('Vehicle not found in config. Skipping: '..k)
-            elseif type(vehicleShop) == 'table' then
-                for i = 1, #vehicleShop do
-                    if vehicleShop[i] == insideShop then
-                        insertVehicle(vehicle)
-                    end
-                end
-            elseif vehicleShop == insideShop then
-                insertVehicle(vehicle)
-            end
-        end
-    end
-
-    table.sort(vehMenu, function(a, b)
-        local _, aName = string.strsplit(' ', string.upper(a.title), 2)
-        local _, bName = string.strsplit(' ', string.upper(b.title), 2)
-
-        return aName < bName
-    end)
+    local vehMenu = require 'client.shop'.loadShop(category, insideShop, targetVehicle)
 
     lib.registerContext({
         id = 'openVehCats',
@@ -533,6 +496,7 @@ local function createShop(shopShape, name)
             insideShop = self.name
         end,
         onExit = function()
+            require 'client.shop'.resetShop()
             insideShop = nil
         end,
     })
